Render a link to the same page on another grafana instance
When migrating from one grafana instance to another, it can be
beneficial to show users that they're on the wrong instance. And if
you did your job right migrating stuff (that is, exporting all your
dashboards and re-importing them with the same IDs in the new place),
you can pretty much just replace some parts of the URL that people
land on, and point them to the next dashboard.

This plugin allows that, with a configurable preamble and link text.